+++ capstone/src/main/java/com/google/univiz/api/CollegeDataApi.java
@@ -1,5 +1,7 @@
-package com.google.univiz;
+package com.google.univiz.api;
 
+import com.google.univiz.api.representation.CollegeData;
+import com.google.univiz.api.representation.CollegeId;
 import java.util.List;
 
 /** Provides access to college related data. */
