--- capstone/src/main/java/com/google/univiz/api/CollegeId.java
@@ -1,8 +1,0 @@
-package com.google.univiz.api;
-
-/**
- * This data structure contains all the college fields that are needed to uniquely identify a
- * college. If two colleges have the same id, they they are the same college. Furthermore this
- * structure may contain other alternative fields that could be used to uniquely identify a college.
- */
-public class CollegeId {}
