+++ capstone/src/main/java/com/google/univiz/api/SuggestionDataApi.java
@@ -1,6 +1,8 @@
-package com.google.univiz;
+package com.google.univiz.api;
 
+import com.google.univiz.api.representation.SuggestionData;
 import com.google.univiz.scorecard.SuggestionResponse;
+import java.util.List;
 
 /**
  * Communicates directly with the data source to get a list of college suggestions related to a
