+++ capstone/src/main/java/com/google/univiz/api/representation/CollegeData.java
@@ -1,4 +1,4 @@
-package com.google.univiz;
+package com.google.univiz.api.representation;
 
 import com.google.auto.value.AutoValue;
