+++ capstone/src/main/java/com/google/univiz/api/representation/SearchData.java
@@ -1,4 +1,4 @@
-package com.google.univiz.api;
+package com.google.univiz.api.representation;
 
 /** SearchData represents the college data specific to the Search feature */
 public class SearchData {}
