+++ capstone/src/main/java/com/google/univiz/api/representation/SuggestionData.java
@@ -1,4 +1,4 @@
-package com.google.univiz;
+package com.google.univiz.api.representation;
 
 /** SuggestionData represents the college data specific to the Search feature */
 public class SuggestionData {}
