+++ capstone/src/main/java/com/google/univiz/api/resource/VisResource.java
@@ -1,8 +1,9 @@
-package com.google.univiz.api;
+package com.google.univiz.api.resource;
 
-import com.google.univiz.api.dataviz.CollegeStats;
-import com.google.univiz.api.dataviz.Deadline;
-import com.google.univiz.api.dataviz.Timeline;
+import com.google.univiz.api.representation.CollegeId;
+import com.google.univiz.api.representation.Deadline;
+import com.google.univiz.api.representation.Timeline;
+import com.google.univiz.api.representation.CollegeStats;
 import java.util.List;
 
 /**
