+++ capstone/src/main/java/com/google/univiz/scorecard/ScorecardConverter.java
@@ -1,9 +1,9 @@
 package com.google.univiz.scorecard;
 
 import com.google.common.base.Converter;
-import com.google.univiz.CarnegieSizeDegree;
-import com.google.univiz.CollegeData;
-import com.google.univiz.CollegeId;
+import com.google.univiz.api.representation.CarnegieSizeDegree;
+import com.google.univiz.api.representation.CollegeData;
+import com.google.univiz.api.representation.CollegeId;
 
 final class ScorecardConverter extends Converter<ScorecardData, CollegeData> {
